Add lint rule to ensure we're using builtins
Since we can now actually check for this more trivially, let's make sure
that we're not requiring things that would would require us to pull from
Galaxy. This adds a new rule similar to the upstream fqcn-builtins rule.
Additionally, the older custom Ansible Lint rule is updated to match the
newer API and to add more documentation.